move tests/util -> tests/__init__.py
