+++ tests/contrib/httplib/test_httplib.py
@@ -16,8 +16,7 @@ from ddtrace.pin import Pin
 
 from tests.opentracer.utils import init_tracer
 
-from ... import TracerTestCase, assert_span_http_status_code
-from ...util import override_global_tracer
+from ... import TracerTestCase, assert_span_http_status_code, override_global_tracer
 
 if PY2:
     from urllib2 import urlopen, build_opener, Request
