+++ tests/contrib/mysqldb/test_mysql.py
@@ -6,8 +6,7 @@ from ddtrace.contrib.mysqldb.patch import patch, unpatch
 
 from tests.opentracer.utils import init_tracer
 from ..config import MYSQL_CONFIG
-from ... import TracerTestCase, assert_is_measured
-from ...util import assert_dict_issuperset
+from ... import TracerTestCase, assert_is_measured, assert_dict_issuperset
 
 
 class MySQLCore(object):
