+++ tests/contrib/pymysql/test_pymysql.py
@@ -10,8 +10,7 @@ from ddtrace.contrib.pymysql.patch import patch, unpatch
 
 # testing
 from tests.opentracer.utils import init_tracer
-from ... import TracerTestCase, assert_is_measured
-from ...util import assert_dict_issuperset
+from ... import TracerTestCase, assert_is_measured, assert_dict_issuperset
 from ...contrib.config import MYSQL_CONFIG
