+++ tests/tracer/test_helpers.py
@@ -2,8 +2,7 @@ import mock
 
 from ddtrace import helpers
 
-from tests import TracerTestCase
-from tests.util import override_global_tracer
+from tests import TracerTestCase, override_global_tracer
 
 
 class HelpersTestCase(TracerTestCase):
